inject self user to linkTextForWholeTeam
